pyproject.toml
[project]
name = "the-evolution-of-todo"
version = "0.1.0"
description = "The Evolution of Todo - From CLI to Full-Stack"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.8.0",
    "sqlmodel>=0.0.14",
    "psycopg2-binary>=2.9.9",
    "pyjwt>=2.8",
    "python-dotenv>=1.0",
    "requests>=2.31",
    "openai>=1.12.0",
    "anthropic>=0.21.0",
    "modelcontextprotocol>=1.0.1",
    "duckduckgo-search>=6.2.6"
]

[tool.uv]
package = false

[tool.uv.workspace]
members = [
    "src/backend",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]
